Disable delete/toggle/edit if task is not a attached to user

+++ src/Controller/TaskController.php
@@ -54,6 +54,10 @@ class TaskController extends AbstractController
      */
     public function edit(Task $task, Request $request, EntityManagerInterface $em): Response
     {
+        if ($task->getUser() !== $this->getUser()) {
+            throw $this->createNotFoundException();
+        }
+
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
@@ -77,6 +81,10 @@ class TaskController extends AbstractController
      */
     public function toggleTask(Task $task, EntityManagerInterface $em): Response
     {
+        if ($task->getUser() !== $this->getUser()) {
+            throw $this->createNotFoundException();
+        }
+
         $task->toggle(!$task->isDone());
         $em->flush();
 
@@ -90,6 +98,10 @@ class TaskController extends AbstractController
      */
     public function deleteTask(Task $task, EntityManagerInterface $em): Response
     {
+        if ($task->getUser() !== $this->getUser()) {
+            throw $this->createNotFoundException();
+        }
+
         $em->remove($task);
         $em->flush();
